refactor: WIP updating Line Account AMOUNTS when unit price or quantity changes [skip ci]
- delete cart CONFIRM popup

+++ libs/react-shared-components/src/lib/cart/edit-cart-table-footer/EditCartTableFooter.tsx
@@ -2,9 +2,25 @@
 // ℹ️ NOT READY YET or NOT MY CODE (chakra templates) ----- please ignore this file, thanks!
 // ##################################################################################
 
-import { Box, Button, Flex, Stat, StatHelpText, StatLabel, StatNumber, TableCaption, Td, Tfoot, Tr } from '@chakra-ui/react';
+import {
+  Box,
+  Button,
+  Flex,
+  Stat,
+  StatHelpText,
+  StatLabel,
+  StatNumber,
+  TableCaption,
+  Td,
+  Tfoot,
+  Tr,
+} from '@chakra-ui/react';
 import { actionIsDeletingCart, store } from '@multi-cart/react-app-state';
-import { Cart, useDeleteCartMutation, useUpdateUserMutation } from '@multi-cart/react-data-access';
+import {
+  Cart,
+  useDeleteCartMutation,
+  useUpdateUserMutation,
+} from '@multi-cart/react-data-access';
 import { sumTotalCost, toFriendlyCurrency } from '@multi-cart/util';
 import { useRouter } from 'next/router';
 import React from 'react';
@@ -12,7 +28,7 @@ import useMyToasts from '../../_hooks/useMyToasts';
 import './EditCartTableFooter.module.scss';
 
 export interface EditCartTableFooterProps {
-  cart: Cart
+  cart: Cart;
 }
 
 export function EditCartTableFooter({ cart }: EditCartTableFooterProps) {
@@ -25,49 +41,62 @@ export function EditCartTableFooter({ cart }: EditCartTableFooterProps) {
     <>
       <TableCaption>
         Prices are estimates and subject to change
-            <Button
-            data-testid="btnDeleteCart"
+        <Button
+          data-testid="btnDeleteCart"
           ml={2}
           size="sm"
           onClick={async () => {
             if (cart?.id) {
-              // TODO: hacky, do we need a setDeleting and unsetDeleting instead of just toggle? or ok... TBD
-              store.dispatch(actionIsDeletingCart); // TOGGLE ON
-              try {
-                const response = await deleteCart({
-                  id: cart.id
-                });
-                if (response.data?.deleteCart === true) {
-                  // and update so NO current cart for this user
-                  const { error, data: updatedUser } = await updateUser({ currentCartId: "" }); // VIP: empty string, just like on iac-side
-                  if (!error && updatedUser?.updateUser?.user?.currentCartId === "") {
-                    toastInfo("Deleted!");
-                    router.push("/dashboard");
-                  } else if (error) {
-                    toastError(error.message);
+              if (
+                window.confirm('Are you SURE you want to 🛑 DELETE this cart?')
+              ) {
+                // TODO: hacky, do we need a setDeleting and unsetDeleting instead of just toggle? or ok... TBD
+                store.dispatch(actionIsDeletingCart); // TOGGLE ON
+                try {
+                  const response = await deleteCart({
+                    id: cart.id,
+                  });
+                  if (response.data?.deleteCart === true) {
+                    // and update so NO current cart for this user
+                    const { error, data: updatedUser } = await updateUser({
+                      currentCartId: '',
+                    }); // VIP: empty string, just like on iac-side
+                    if (
+                      !error &&
+                      updatedUser?.updateUser?.user?.currentCartId === ''
+                    ) {
+                      toastInfo('Deleted!');
+                      router.push('/dashboard');
+                    } else if (error) {
+                      toastError(error.message);
+                    }
                   }
+                } finally {
+                  setTimeout(() => store.dispatch(actionIsDeletingCart), 2000); // TOGGLE OFF (with delay)
                 }
-              } finally {
-                setTimeout(() => store.dispatch(actionIsDeletingCart), 2000);  // TOGGLE OFF (with delay)
               }
             }
-          }}>
+          }}
+        >
           <strong>Delete</strong> Cart
-              </Button>
+        </Button>
       </TableCaption>
       <Tfoot>
         <Tr>
-          <Td
-            align="right"
-            colSpan={28}>
+          <Td align="right" colSpan={28}>
             <Flex
               justifyContent="flex-end"
-              mr={{ base: -3, md: -3, xl: "75px" }}>
+              mr={{ base: -3, md: -3, xl: '75px' }}
+            >
               <Box>
                 <Stat align="right">
                   <StatLabel>Total:</StatLabel>
-                  <StatNumber>{toFriendlyCurrency(sumTotalCost(cart))}</StatNumber>
-                  <StatHelpText fontWeight="hairline">not including shipping</StatHelpText>
+                  <StatNumber>
+                    {toFriendlyCurrency(sumTotalCost(cart))}
+                  </StatNumber>
+                  <StatHelpText fontWeight="hairline">
+                    not including shipping
+                  </StatHelpText>
                 </Stat>
               </Box>
             </Flex>
